+++ deleteManager.js
@@ -1,5 +1,4 @@
 // version 1.0.0 - FAANG-Level Refactored
-import tools from './Tools.js';
 
 // Constants for relationship types
 const RELATIONSHIP_TYPES = {